requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
idna==3.6
mypy-protobuf==3.5.0
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
types-protobuf==4.24.0.4
typing_extensions==4.8.0
uvicorn==0.24.0.post1
